Update nsm criteria.

+++ lattice_quantizer/criteria/baseline.py
@@ -1,10 +1,10 @@
 import numpy as np
 
 # 1D: Integer lattice Z
-B_Z = np.array([[1.0]])
+B_Z = np.array([[1.0]], dtype=float)
 
 # 2D: A2 lattice
-B_A2 = np.array([[1.0, 0.0], [-0.5, np.sqrt(3) / 2]])
+B_A2 = np.array([[1.0, 0.0], [-0.5, np.sqrt(3) / 2]], dtype=float)
 
 # 3D: FCC (D3) lattice
 B_D3 = np.array(
@@ -12,7 +12,8 @@ B_D3 = np.array(
         [1.0, 0.0, 0.0],
         [-0.5, np.sqrt(3) / 2, 0.0],
         [-0.5, -np.sqrt(3) / 6, np.sqrt(2 / 3)],
-    ]
+    ],
+    dtype=float,
 )
 
 # 4D: D4 lattice
@@ -22,5 +23,6 @@ B_D4 = np.array(
         [-0.5, np.sqrt(3) / 2, 0.0, 0.0],
         [-0.5, -np.sqrt(3) / 6, np.sqrt(2 / 3), 0.0],
         [-0.5, -np.sqrt(3) / 6, -np.sqrt(2) / 3, np.sqrt(5 / 6)],
-    ]
+    ],
+    dtype=float,
 )
